pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scio"
version = "0.1.0"
description = "SCIO - Scientific Intelligent Operations Framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "SCIO Team" }
]
keywords = [
    "science",
    "agents",
    "reproducibility",
    "experiments",
    "automation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "structlog>=23.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]
data = [
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
]
all = [
    "scio[api,llm,data]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "black>=23.0",
    "httpx>=0.26.0",
]

[project.scripts]
scio = "scio.cli.main:app"

[project.urls]
Homepage = "https://github.com/scio-framework/scio"
Documentation = "https://scio-framework.github.io/scio"
Repository = "https://github.com/scio-framework/scio"

[tool.hatch.build.targets.wheel]
packages = ["scio"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.isort]
known-first-party = ["scio"]

[tool.black]
line-length = 100
target-version = ["py311"]
